refactor(tests): split test_sync_engine.py per #275 pattern (#570)
services/saves/sync_engine/ was decomposed into engine/matrix/devices/
rollback in #495. The test file stayed a 2200+ LOC monolith. Splits
into tests/services/saves/sync_engine/test_{engine,matrix,devices,rollback}.py
mirroring the source layout per the #275 convention already applied
elsewhere in tests/services/saves/.

Pure structural move. Test count unchanged. No logic changes.

Closes #570
