+++ tests/services/saves/sync_engine/test_matrix.py
@@ -1,12 +1,17 @@
-"""Tests for SyncEngine — newest-wins sync matrix and concurrent-sync coordination."""
+"""Tests for MatrixExecutor — newest-wins matrix evaluation and per-file sync I/O
+dispatch. Anything that decides "which side wins for this file" or moves bytes
+between local saves_dir and the RomM server lives in
+py_modules/services/saves/sync_engine/matrix.py and is exercised here. Public-
+callable orchestration (lock acquisition, guards) lives in test_engine.py;
+device registration in test_devices.py; conflict rollback in test_rollback.py.
+"""
 
 import hashlib
-import logging
 import os
 
 import pytest
 
-from domain.save_state import FileSyncState, RomSaveState
+from domain.save_state import RomSaveState
 from lib.errors import RommApiError
 from tests.services.saves._helpers import (
     _create_save,
@@ -17,518 +22,6 @@ from tests.services.saves._helpers import (
     _server_save_with_syncs,
     make_service,
 )
-
-
-class TestSyncRomSaves:
-    def test_local_only_uploads(self, tmp_path):
-        svc, fake = make_service(tmp_path)
-        svc._save_sync_state.settings.save_sync_enabled = True
-        _install_rom(svc, tmp_path)
-        _create_save(tmp_path, content=b"save data")
-
-        synced, errors, conflicts = svc._sync_engine._sync_rom_saves(42)
-        assert synced == 1
-        assert errors == []
-        assert conflicts == []
-        assert any(c[0] == "upload_save" for c in fake.call_log)
-
-    def test_server_only_downloads(self, tmp_path):
-        svc, fake = make_service(tmp_path)
-        svc._save_sync_state.settings.save_sync_enabled = True
-        _install_rom(svc, tmp_path)
-        # Add server save but no local file
-        ss = _server_save()
-        fake.saves[100] = ss
-
-        synced, errors, _ = svc._sync_engine._sync_rom_saves(42)
-        assert synced == 1
-        assert errors == []
-        # Verify the file was downloaded
-        saves_dir = tmp_path / "saves" / "gba"
-        assert (saves_dir / "pokemon.srm").exists()
-
-    def test_rom_not_installed(self, tmp_path):
-        svc, _ = make_service(tmp_path)
-        synced, errors, _ = svc._sync_engine._sync_rom_saves(999)
-        assert synced == 0
-        assert errors == []
-
-    def test_api_error_on_list_saves(self, tmp_path):
-        svc, fake = make_service(tmp_path)
-        _install_rom(svc, tmp_path)
-        fake.fail_on_next(RommApiError("Server error"))
-
-        synced, errors, _ = svc._sync_engine._sync_rom_saves(42)
-        assert synced == 0
-        assert len(errors) == 1
-        assert "Failed to fetch saves" in errors[0]
-
-    # ------------------------------------------------------------------
-    # Regression tests for issue #238 — pending-migration handling.
-    # Rule 2: skip server_only downloads while a save-sort migration is
-    # pending so the mtime-naive resolver cannot prefer freshly-downloaded
-    # stale server content over real user progress at the other layout.
-    # ------------------------------------------------------------------
-
-    def test_sync_rom_saves_skips_server_only_downloads_during_pending_migration(self, tmp_path):
-        """server_only matches must be skipped while migration is pending (#238)."""
-        svc, fake = make_service(tmp_path)
-        svc._save_sync_state.settings.save_sync_enabled = True
-        _install_rom(svc, tmp_path)
-        # Mark migration pending — detect has fired, user hasn't resolved yet.
-        svc._state["save_sort_settings"] = {"sort_by_content": True, "sort_by_core": False}
-        svc._state["save_sort_settings_previous"] = {"sort_by_content": True, "sort_by_core": False}
-        # Server has a save, no local file anywhere.
-        ss = _server_save()
-        fake.saves[100] = ss
-
-        synced, errors, conflicts = svc._sync_engine._sync_rom_saves(42)
-
-        assert synced == 0
-        assert errors == []
-        assert conflicts == []
-        # No download was initiated.
-        assert fake.downloaded_files == {}
-        # No file landed on disk under either layout.
-        saves_dir = tmp_path / "saves" / "gba"
-        assert not (saves_dir / "pokemon.srm").exists()
-
-    def test_sync_rom_saves_uploads_local_only_during_pending_migration(self, tmp_path):
-        """local_only matches must still upload during pending migration (#238)."""
-        svc, fake = make_service(tmp_path)
-        svc._save_sync_state.settings.save_sync_enabled = True
-        _install_rom(svc, tmp_path)
-        svc._state["save_sort_settings"] = {"sort_by_content": True, "sort_by_core": False}
-        svc._state["save_sort_settings_previous"] = {"sort_by_content": True, "sort_by_core": False}
-        # Local save at the (previous == current, same layout) location.
-        _create_save(tmp_path, content=b"user progress")
-
-        synced, errors, conflicts = svc._sync_engine._sync_rom_saves(42)
-
-        assert synced == 1
-        assert errors == []
-        assert conflicts == []
-        # Upload went through.
-        assert any(c[0] == "upload_save" for c in fake.call_log)
-
-    @pytest.mark.asyncio
-    async def test_sync_rom_saves_invokes_detect_sort_change_before_sync(self, tmp_path):
-        """Manual sync_rom_saves must also refresh save-sort state first (#238).
-
-        Without the detect-first call, a user editing retroarch.cfg outside
-        of a session and then triggering manual sync would race the same
-        way that direct-Steam-launch does — sync would compute saves_dir
-        from stale state and risk landing stale server content at the
-        wrong layout.
-        """
-        call_order: list[str] = []
-
-        def fake_detect() -> None:
-            call_order.append("detect")
-
-        svc, _ = make_service(tmp_path, detect_sort_change=fake_detect)
-        svc._save_sync_state.settings.save_sync_enabled = True
-        svc._save_sync_state.device_id = "test-device"
-        _install_rom(svc, tmp_path)
-        _create_save(tmp_path, content=b"progress")
-
-        orig_sync = svc._sync_engine._sync_rom_saves
-
-        def wrapped_sync(rom_id):
-            call_order.append("sync")
-            return orig_sync(rom_id)
-
-        svc._sync_engine._sync_rom_saves = wrapped_sync  # type: ignore[method-assign]
-
-        result = await svc.sync_rom_saves(42)
-
-        assert result["success"] is True
-        # detect fired exactly once, before sync ran.
-        assert call_order.count("detect") == 1
-        assert call_order.index("detect") < call_order.index("sync")
-
-    @pytest.mark.asyncio
-    async def test_sync_rom_saves_message_includes_conflict_count(self, tmp_path):
-        """Public sync_rom_saves must surface conflict count in its message.
-
-        Previously reported "Synced 0 save(s)" even with conflicts present,
-        which reads as success — user had no signal that manual intervention
-        was needed.
-        """
-        svc, _ = make_service(tmp_path)
-        svc._save_sync_state.settings.save_sync_enabled = True
-        svc._save_sync_state.device_id = "test-device"
-        _install_rom(svc, tmp_path)
-
-        # Stub _sync_rom_saves to return 1 conflict, 0 synced, 0 errors
-        def stub_sync(rom_id):
-            return (0, [], [{"type": "newer_in_slot", "rom_id": rom_id}])
-
-        svc._sync_engine._sync_rom_saves = stub_sync  # type: ignore[method-assign]
-
-        result = await svc.sync_rom_saves(42)
-
-        # success is still True — conflicts are legitimate state, not technical failure
-        assert result["success"] is True
-        assert "1 conflict(s)" in result["message"]
-        assert result["synced"] == 0
-
-
-class TestSyncAllSaves:
-    @pytest.mark.asyncio
-    async def test_syncs_multiple_roms(self, tmp_path):
-        svc, _ = make_service(tmp_path)
-        svc._save_sync_state.settings.save_sync_enabled = True
-        svc._save_sync_state.device_id = "test-device"
-
-        _install_rom(svc, tmp_path, rom_id=1, system="gba", file_name="game1.gba")
-        _install_rom(svc, tmp_path, rom_id=2, system="snes", file_name="game2.sfc")
-        _create_save(tmp_path, system="gba", rom_name="game1", content=b"save1")
-        _create_save(tmp_path, system="snes", rom_name="game2", content=b"save2")
-
-        result = await svc.sync_all_saves()
-        assert result["success"] is True
-        assert result["synced"] == 2
-        assert result["roms_checked"] == 2
-
-    @pytest.mark.asyncio
-    async def test_disabled_returns_early(self, tmp_path):
-        svc, _ = make_service(tmp_path)
-        result = await svc.sync_all_saves()
-        assert result["success"] is False
-        assert "disabled" in result["message"].lower()
-
-    @pytest.mark.asyncio
-    async def test_partial_failure(self, tmp_path):
-        svc, fake = make_service(tmp_path)
-        svc._save_sync_state.settings.save_sync_enabled = True
-        svc._save_sync_state.device_id = "test-device"
-
-        _install_rom(svc, tmp_path, rom_id=1, system="gba", file_name="game1.gba")
-        _install_rom(svc, tmp_path, rom_id=2, system="snes", file_name="game2.sfc")
-        _create_save(tmp_path, system="gba", rom_name="game1", content=b"save1")
-        _create_save(tmp_path, system="snes", rom_name="game2", content=b"save2")
-
-        # Make the second ROM's list_saves fail
-        original_list = fake.list_saves
-
-        call_count = 0
-
-        def flaky_list(rom_id, *, device_id=None, slot=None):
-            nonlocal call_count
-            call_count += 1
-            if call_count == 2:
-                raise RommApiError("Server error")
-            return original_list(rom_id, device_id=device_id, slot=slot)
-
-        fake.list_saves = flaky_list
-
-        result = await svc.sync_all_saves()
-        assert result["synced"] >= 1
-        assert len(result["errors"]) >= 1
-
-    @pytest.mark.asyncio
-    async def test_sync_all_saves_invokes_detect_sort_change_before_sync(self, tmp_path):
-        """Manual sync_all_saves must also refresh save-sort state first (#238).
-
-        Same race as sync_rom_saves but for the bulk path: detect must
-        fire once at the top of the method, before any per-ROM sync runs.
-        """
-        call_order: list[str] = []
-
-        def fake_detect() -> None:
-            call_order.append("detect")
-
-        svc, _ = make_service(tmp_path, detect_sort_change=fake_detect)
-        svc._save_sync_state.settings.save_sync_enabled = True
-        svc._save_sync_state.device_id = "test-device"
-        _install_rom(svc, tmp_path, rom_id=1, system="gba", file_name="game1.gba")
-        _create_save(tmp_path, system="gba", rom_name="game1", content=b"save1")
-
-        orig_sync = svc._sync_engine._sync_rom_saves
-
-        def wrapped_sync(rom_id):
-            call_order.append("sync")
-            return orig_sync(rom_id)
-
-        svc._sync_engine._sync_rom_saves = wrapped_sync  # type: ignore[method-assign]
-
-        result = await svc.sync_all_saves()
-
-        assert result["success"] is True
-        # detect fired exactly once, before any per-ROM sync ran.
-        assert call_order.count("detect") == 1
-        assert call_order.index("detect") < call_order.index("sync")
-
-    @pytest.mark.asyncio
-    async def test_sync_all_saves_success_stays_true_with_only_conflicts(self, tmp_path):
-        """Regression guard: success flag reflects errors only, not conflicts.
-
-        Conflicts are a legitimate state requiring user resolution — not a
-        technical failure. Frontend distinguishes via conflicts count; success
-        flag must stay reserved for actual errors.
-        """
-        svc, _ = make_service(tmp_path)
-        svc._save_sync_state.settings.save_sync_enabled = True
-        svc._save_sync_state.device_id = "test-device"
-        _install_rom(svc, tmp_path, rom_id=1, system="gba", file_name="game1.gba")
-
-        # Stub internal sync to produce conflicts but no errors
-        def stub_sync(rom_id):
-            return (0, [], [{"type": "newer_in_slot", "rom_id": rom_id}])
-
-        svc._sync_engine._sync_rom_saves = stub_sync  # type: ignore[method-assign]
-
-        result = await svc.sync_all_saves()
-
-        assert result["success"] is True
-        assert result["conflicts"] >= 1
-        assert "conflict(s)" in result["message"]
-
-
-class TestPreLaunchSync:
-    @pytest.mark.asyncio
-    async def test_downloads_server_saves(self, tmp_path):
-        svc, fake = make_service(tmp_path)
-        svc._save_sync_state.settings.save_sync_enabled = True
-        svc._save_sync_state.device_id = "test-device"
-        _install_rom(svc, tmp_path)
-
-        ss = _server_save()
-        fake.saves[100] = ss
-
-        result = await svc.pre_launch_sync(42)
-        assert result["success"] is True
-        assert result["synced"] == 1
-
-    @pytest.mark.asyncio
-    async def test_disabled_skips(self, tmp_path):
-        svc, _ = make_service(tmp_path)
-        result = await svc.pre_launch_sync(42)
-        assert result["success"] is True
-        assert result["synced"] == 0
-
-    @pytest.mark.asyncio
-    async def test_pre_launch_disabled_in_settings(self, tmp_path):
-        svc, _ = make_service(tmp_path)
-        svc._save_sync_state.settings.save_sync_enabled = True
-        svc._save_sync_state.settings.sync_before_launch = False
-        svc._save_sync_state.device_id = "test-device"
-
-        result = await svc.pre_launch_sync(42)
-        assert result["synced"] == 0
-
-    @pytest.mark.asyncio
-    async def test_pre_launch_sync_invokes_detect_sort_change_before_migration_gate(self, tmp_path):
-        """detect_sort_change is called before the _is_save_sort_changed gate (#238)."""
-        order: list[str] = []
-
-        def fake_detect() -> None:
-            # Simulate detect discovering a pending migration.
-            order.append("detect")
-
-        svc, _ = make_service(tmp_path, detect_sort_change=fake_detect)
-        svc._save_sync_state.settings.save_sync_enabled = True
-        svc._save_sync_state.device_id = "test-device"
-
-        # Track when is_save_sort_changed is consulted.
-        orig_gate = svc._rom_info.is_save_sort_changed
-
-        def wrapped_gate():
-            order.append("gate")
-            return orig_gate()
-
-        svc._rom_info.is_save_sort_changed = wrapped_gate  # type: ignore[method-assign]
-
-        await svc.pre_launch_sync(42)
-
-        assert "detect" in order
-        assert "gate" in order
-        assert order.index("detect") < order.index("gate")
-
-
-class TestPostExitSync:
-    @pytest.mark.asyncio
-    async def test_uploads_changed_saves(self, tmp_path):
-        svc, _ = make_service(tmp_path)
-        svc._save_sync_state.settings.save_sync_enabled = True
-        svc._save_sync_state.device_id = "test-device"
-        _install_rom(svc, tmp_path)
-        _create_save(tmp_path, content=b"new save data")
-
-        result = await svc.post_exit_sync(42)
-        assert result["success"] is True
-        assert result["synced"] == 1
-
-    @pytest.mark.asyncio
-    async def test_disabled_skips(self, tmp_path):
-        svc, _ = make_service(tmp_path)
-        result = await svc.post_exit_sync(42)
-        assert result["success"] is True
-        assert result["synced"] == 0
-
-    @pytest.mark.asyncio
-    async def test_post_exit_disabled_in_settings(self, tmp_path):
-        svc, _ = make_service(tmp_path)
-        svc._save_sync_state.settings.save_sync_enabled = True
-        svc._save_sync_state.settings.sync_after_exit = False
-        svc._save_sync_state.device_id = "test-device"
-
-        result = await svc.post_exit_sync(42)
-        assert result["synced"] == 0
-
-    @pytest.mark.asyncio
-    async def test_auto_registers_device(self, tmp_path):
-        svc, _ = make_service(tmp_path)
-        svc._save_sync_state.settings.save_sync_enabled = True
-        # No device_id set
-        _install_rom(svc, tmp_path)
-        _create_save(tmp_path, content=b"data")
-
-        result = await svc.post_exit_sync(42)
-        assert result["success"] is True
-        assert svc._save_sync_state.device_id is not None
-
-    # ------------------------------------------------------------------
-    # Regression tests for issue #238 — detect-first invariant.
-    #
-    # Save-sync must refresh save-sort state via detect_sort_change
-    # before computing saves_dir, so that Rule 1 / Rule 2 engage even
-    # when a direct-Steam-launch race delivers post_exit_sync before
-    # refreshMigrationState. See #238.
-    # ------------------------------------------------------------------
-
-    @pytest.mark.asyncio
-    async def test_post_exit_sync_invokes_detect_sort_change_before_sync(self, tmp_path):
-        """detect_sort_change is called exactly once before the sync path runs (#238)."""
-        call_order: list[str] = []
-
-        def fake_detect() -> None:
-            call_order.append("detect")
-
-        svc, _ = make_service(tmp_path, detect_sort_change=fake_detect)
-        svc._save_sync_state.settings.save_sync_enabled = True
-        svc._save_sync_state.device_id = "test-device"
-        _install_rom(svc, tmp_path)
-        _create_save(tmp_path, content=b"progress")
-
-        # Patch _sync_rom_saves to record call ordering.
-        orig_sync = svc._sync_engine._sync_rom_saves
-
-        def wrapped_sync(rom_id):
-            call_order.append("sync")
-            return orig_sync(rom_id)
-
-        svc._sync_engine._sync_rom_saves = wrapped_sync  # type: ignore[method-assign]
-
-        result = await svc.post_exit_sync(42)
-
-        assert result["success"] is True
-        # detect fired exactly once, before sync ran.
-        assert call_order.count("detect") == 1
-        assert call_order.index("detect") < call_order.index("sync")
-
-    @pytest.mark.asyncio
-    async def test_post_exit_sync_continues_when_detect_sort_change_raises(self, tmp_path, caplog):
-        """If detect_sort_change raises, save-sync logs a warning and proceeds (#238)."""
-
-        def boom() -> None:
-            raise RuntimeError("cfg file unreadable")
-
-        svc, _ = make_service(tmp_path, detect_sort_change=boom)
-        svc._save_sync_state.settings.save_sync_enabled = True
-        svc._save_sync_state.device_id = "test-device"
-        _install_rom(svc, tmp_path)
-        _create_save(tmp_path, content=b"progress")
-
-        with caplog.at_level(logging.WARNING, logger="test"):
-            result = await svc.post_exit_sync(42)
-
-        assert result["success"] is True
-        # Sync still ran despite detect failure.
-        assert result["synced"] == 1
-        # Warning was logged.
-        assert any("detect_sort_change failed" in rec.message for rec in caplog.records)
-
-    @pytest.mark.asyncio
-    async def test_post_exit_sync_works_when_detect_sort_change_is_none(self, tmp_path):
-        """Default detect_sort_change=None: post_exit_sync still runs without error (#238)."""
-        svc, _ = make_service(tmp_path)  # detect_sort_change not passed → None
-        assert svc._sync_engine._detect_sort_change is None
-        svc._save_sync_state.settings.save_sync_enabled = True
-        svc._save_sync_state.device_id = "test-device"
-        _install_rom(svc, tmp_path)
-        _create_save(tmp_path, content=b"progress")
-
-        result = await svc.post_exit_sync(42)
-
-        assert result["success"] is True
-        assert result["synced"] == 1
-
-    @pytest.mark.asyncio
-    async def test_post_exit_sync_message_includes_conflict_count(self, tmp_path):
-        """post_exit_sync must surface conflict count in its message.
-
-        Previously "Uploaded 0 save(s)" even with conflicts — user has no
-        signal that sync is blocked on manual resolution.
-        """
-        svc, _ = make_service(tmp_path)
-        svc._save_sync_state.settings.save_sync_enabled = True
-        svc._save_sync_state.device_id = "test-device"
-        _install_rom(svc, tmp_path)
-
-        def stub_sync(rom_id):
-            return (0, [], [{"type": "newer_in_slot", "rom_id": rom_id}])
-
-        svc._sync_engine._sync_rom_saves = stub_sync  # type: ignore[method-assign]
-
-        result = await svc.post_exit_sync(42)
-
-        assert result["success"] is True
-        assert "1 conflict(s)" in result["message"]
-        assert result["synced"] == 0
-
-
-class TestCheckSaveStatusBackground:
-    """Tests for the background save status check with event emit."""
-
-    @pytest.mark.asyncio
-    async def test_emits_save_status_updated(self, tmp_path):
-        """Background check runs full status and emits result."""
-        emitted = []
-
-        async def fake_emit(event, *args):
-            emitted.append((event, args))
-
-        svc, _fake = make_service(tmp_path, emit=fake_emit)
-        _install_rom(svc, tmp_path)
-        _create_save(tmp_path)
-
-        await svc.check_save_status_background(42)
-
-        assert len(emitted) == 1
-        assert emitted[0][0] == "save_status_updated"
-        result = emitted[0][1][0]
-        assert result["rom_id"] == 42
-        assert len(result["files"]) >= 1
-
-    @pytest.mark.asyncio
-    async def test_no_emit_when_emit_is_none(self, tmp_path):
-        """Background check works without emit (no crash)."""
-        svc, _ = make_service(tmp_path)
-        _install_rom(svc, tmp_path)
-
-        # Should not raise even without emit
-        await svc.check_save_status_background(42)
-
-    @pytest.mark.asyncio
-    async def test_swallows_errors(self, tmp_path):
-        """Background check logs but does not raise on errors."""
-        svc, fake = make_service(tmp_path)
-        fake.fail_on_next(Exception("Server down"))
-
-        # Should not raise
-        await svc.check_save_status_background(42)
 
 
 class TestUploadSpecialChars:
@@ -1745,752 +1238,6 @@ class TestSyncRomSavesDispatch:
         assert after is not None and isinstance(after, str)
 
 
-class TestResolveSyncConflict:
-    @pytest.mark.asyncio
-    async def test_resolve_keep_local_hash_match_short_circuits(self, tmp_path):
-        """Local hash matches server's content hash → no PUT, state updated."""
-        svc, fake = make_service(tmp_path)
-        _enable_sync_with_device(svc)
-        _install_rom(svc, tmp_path)
-        save_path = _create_save(tmp_path, content=b"identical content")
-        local_hash = _file_md5(str(save_path))
-
-        ss = _server_save_with_syncs(
-            device_syncs=[{"device_id": "device-1", "is_current": False}],
-        )
-        # Make the server hash equal to the local hash by uploading the same
-        # file as the source: FakeSaveApi.download_save copies uploaded_files.
-        fake.saves[100] = ss
-        fake.uploaded_files[100] = str(save_path)
-
-        result = await svc.resolve_sync_conflict(
-            rom_id=42,
-            filename="pokemon.srm",
-            server_save_id=100,
-            action="keep_local",
-        )
-
-        assert result["success"] is True
-        assert result["action"] == "keep_local"
-        assert not any(c[0] == "upload_save" for c in fake.call_log)
-
-        file_state = svc._save_sync_state.saves["42"].files["pokemon.srm"]
-        assert file_state.tracked_save_id == 100
-        assert file_state.last_sync_hash == local_hash
-
-    @pytest.mark.asyncio
-    async def test_resolve_keep_local_hash_mismatch_uploads_put(self, tmp_path):
-        """Local differs from server content → PUT against existing save id."""
-        svc, fake = make_service(tmp_path)
-        _enable_sync_with_device(svc)
-        _install_rom(svc, tmp_path)
-        save_path = _create_save(tmp_path, content=b"local-edited")
-        local_hash = _file_md5(str(save_path))
-
-        ss = _server_save_with_syncs(
-            device_syncs=[{"device_id": "device-1", "is_current": False}],
-        )
-        # Server has different content uploaded — hash will not match.
-        other = tmp_path / "other.bin"
-        other.write_bytes(b"server-flavor")
-        fake.saves[100] = ss
-        fake.uploaded_files[100] = str(other)
-
-        result = await svc.resolve_sync_conflict(
-            rom_id=42,
-            filename="pokemon.srm",
-            server_save_id=100,
-            action="keep_local",
-        )
-
-        assert result["success"] is True
-        upload_calls = [c for c in fake.call_log if c[0] == "upload_save"]
-        assert len(upload_calls) == 1
-        # PUT — save_id was passed
-        assert upload_calls[0][2]["save_id"] == 100
-
-        file_state = svc._save_sync_state.saves["42"].files["pokemon.srm"]
-        assert file_state.last_sync_hash == local_hash
-
-    @pytest.mark.asyncio
-    async def test_resolve_keep_local_falls_back_when_server_hash_fetch_raises(self, tmp_path):
-        """When ``_get_server_save_hash`` raises out of retry (retries exhausted),
-        the keep_local resolver swallows it and treats ``server_hash`` as ``None``.
-
-        That sinks the adopt-without-upload short-circuit (which requires
-        ``server_hash`` truthy and equal to ``local_hash``) and the PUT path
-        runs unconditionally. Degraded behaviour, but no failure surfaced to
-        the user.
-        """
-        svc, fake = make_service(tmp_path)
-        _enable_sync_with_device(svc)
-        _install_rom(svc, tmp_path)
-        save_path = _create_save(tmp_path, content=b"local-edited")
-        local_hash = _file_md5(str(save_path))
-
-        ss = _server_save_with_syncs(
-            device_syncs=[{"device_id": "device-1", "is_current": False}],
-        )
-        fake.saves[100] = ss
-        # No uploaded_files entry — wouldn't matter anyway, the download_save
-        # call inside _get_server_save_hash is going to raise before reading.
-
-        # Make ``_get_server_save_hash`` re-raise: ``download_save`` raises
-        # and the retry mock reports the exception as retryable, so the
-        # inner ``except Exception`` in ``_get_server_save_hash`` re-raises
-        # and the outer ``except Exception`` in
-        # ``_resolve_conflict_keep_local`` catches it. We monkey-patch
-        # ``download_save`` directly (rather than ``fail_on_next``, which
-        # would consume on the earlier ``list_saves`` call).
-        def _raise_on_download(save_id: int, dest_path: str) -> None:
-            fake.call_log.append(("download_save", (save_id, dest_path), {}))
-            raise RommApiError("transient")
-
-        fake.download_save = _raise_on_download  # type: ignore[method-assign]
-        svc._sync_engine._retry.is_retryable.return_value = True  # type: ignore[attr-defined]
-
-        result = await svc.resolve_sync_conflict(
-            rom_id=42,
-            filename="pokemon.srm",
-            server_save_id=100,
-            action="keep_local",
-        )
-
-        assert result["success"] is True
-        assert result["action"] == "keep_local"
-        # The hash-match short-circuit MUST NOT have fired — its branch
-        # records ``tracked_save_id`` without an upload. We instead expect
-        # the PUT upload path to have run.
-        upload_calls = [c for c in fake.call_log if c[0] == "upload_save"]
-        assert len(upload_calls) == 1
-        # PUT — save_id was passed (existing save id, not None).
-        assert upload_calls[0][2]["save_id"] == 100
-        # download_save was attempted exactly once (the one that raised).
-        download_calls = [c for c in fake.call_log if c[0] == "download_save"]
-        assert len(download_calls) == 1
-
-        # State carries the local hash from the successful PUT.
-        file_state = svc._save_sync_state.saves["42"].files["pokemon.srm"]
-        assert file_state.last_sync_hash == local_hash
-
-    @pytest.mark.asyncio
-    async def test_resolve_use_server_downloads_and_persists(self, tmp_path):
-        """use_server downloads server, overwrites local, updates state."""
-        svc, fake = make_service(tmp_path)
-        _enable_sync_with_device(svc)
-        _install_rom(svc, tmp_path)
-        save_path = _create_save(tmp_path, content=b"local-stale")
-
-        # Server has different content
-        server_content = tmp_path / "server-content.bin"
-        server_content.write_bytes(b"server-truth")
-        ss = _server_save_with_syncs(
-            device_syncs=[{"device_id": "device-1", "is_current": False}],
-        )
-        fake.saves[100] = ss
-        fake.uploaded_files[100] = str(server_content)
-
-        result = await svc.resolve_sync_conflict(
-            rom_id=42,
-            filename="pokemon.srm",
-            server_save_id=100,
-            action="use_server",
-        )
-
-        assert result["success"] is True
-        # Local file overwritten with server content
-        assert save_path.read_bytes() == b"server-truth"
-        file_state = svc._save_sync_state.saves["42"].files["pokemon.srm"]
-        assert file_state.tracked_save_id == 100
-        assert file_state.last_sync_hash == _file_md5(str(save_path))
-
-    @pytest.mark.asyncio
-    async def test_resolve_invalid_action_returns_error(self, tmp_path):
-        svc, _ = make_service(tmp_path)
-        _enable_sync_with_device(svc)
-        _install_rom(svc, tmp_path)
-
-        result = await svc.resolve_sync_conflict(
-            rom_id=42,
-            filename="pokemon.srm",
-            server_save_id=100,
-            action="foo",
-        )
-
-        assert result["success"] is False
-        assert "invalid" in result["message"].lower()
-
-    @pytest.mark.asyncio
-    async def test_resolve_rom_not_installed(self, tmp_path):
-        svc, _ = make_service(tmp_path)
-        _enable_sync_with_device(svc)
-
-        result = await svc.resolve_sync_conflict(
-            rom_id=999,
-            filename="pokemon.srm",
-            server_save_id=100,
-            action="keep_local",
-        )
-
-        assert result["success"] is False
-        assert result["message"]
-
-    @pytest.mark.asyncio
-    async def test_resolve_server_fetch_failure(self, tmp_path):
-        """When list_saves raises, return failure without mutating state."""
-        svc, fake = make_service(tmp_path)
-        _enable_sync_with_device(svc)
-        _install_rom(svc, tmp_path)
-        _create_save(tmp_path, content=b"x")
-
-        # Pre-populate state to assert it stays untouched.
-        original_state = RomSaveState(
-            files={"pokemon.srm": FileSyncState(tracked_save_id=100, last_sync_hash="abc")},
-        )
-        svc._save_sync_state.saves["42"] = original_state
-
-        fake.fail_on_next(RommApiError("network"))
-
-        result = await svc.resolve_sync_conflict(
-            rom_id=42,
-            filename="pokemon.srm",
-            server_save_id=100,
-            action="keep_local",
-        )
-
-        assert result["success"] is False
-        assert "Failed to fetch saves" in result["message"]
-        # State left as-is — no mutation
-        assert svc._save_sync_state.saves["42"] == original_state
-
-    @pytest.mark.asyncio
-    async def test_resolve_no_server_saves_in_slot(self, tmp_path):
-        """Empty slot post-fetch returns success=False with a clear message.
-
-        Implementation note: ``resolve_sync_conflict`` reaches the slot-empty
-        branch via ``_filter_server_saves_to_slot`` and returns
-        ``{"success": False, "message": "No server save in active slot"}``.
-        """
-        svc, _ = make_service(tmp_path)
-        _enable_sync_with_device(svc)
-        _install_rom(svc, tmp_path)
-
-        result = await svc.resolve_sync_conflict(
-            rom_id=42,
-            filename="pokemon.srm",
-            server_save_id=100,
-            action="keep_local",
-        )
-
-        assert result["success"] is False
-        assert "no server save" in result["message"].lower()
-
-    @pytest.mark.asyncio
-    async def test_resolve_filename_symmetry_keep_local_uses_canonical_target(self, tmp_path):
-        """keep_local and use_server must resolve the on-disk path the same way.
-
-        Both branches must derive ``<rom_name>.<server.file_extension>`` from
-        the server save, ignoring the frontend-supplied ``filename`` for I/O.
-        Otherwise an extension drift between the frontend label and the
-        canonical name produces divergent disk and state outcomes for the
-        same conflict.
-        """
-        svc, fake = make_service(tmp_path)
-        _enable_sync_with_device(svc)
-        _install_rom(svc, tmp_path)
-        # Canonical local save: pokemon.srm (matches server.file_extension).
-        canonical_path = _create_save(tmp_path, content=b"local-progress")
-        canonical_hash = _file_md5(str(canonical_path))
-
-        # Server save advertises file_extension=srm; frontend will send a
-        # mismatched filename (pokemon.sav). Server hash differs so the
-        # adopt-without-upload short-circuit doesn't fire.
-        ss = _server_save_with_syncs(
-            device_syncs=[{"device_id": "device-1", "is_current": False}],
-        )
-        ss["file_extension"] = "srm"
-        other = tmp_path / "server-bytes.bin"
-        other.write_bytes(b"server-flavor")
-        fake.saves[100] = ss
-        fake.uploaded_files[100] = str(other)
-
-        result = await svc.resolve_sync_conflict(
-            rom_id=42,
-            filename="pokemon.sav",  # diverges from server canonical
-            server_save_id=100,
-            action="keep_local",
-        )
-
-        assert result["success"] is True
-        # Canonical local file remained — no rename, no orphan at the
-        # frontend-supplied name.
-        assert canonical_path.read_bytes() == b"local-progress"
-        assert not (tmp_path / "saves" / "gba" / "pokemon.sav").exists()
-        # Upload PUT used the canonical filename, not the user-supplied one.
-        upload_calls = [c for c in fake.call_log if c[0] == "upload_save"]
-        assert len(upload_calls) == 1
-        upload_path = upload_calls[0][1][1]
-        assert os.path.basename(upload_path) == "pokemon.srm"
-        # State keyed by canonical filename — never by the frontend label.
-        files_state = svc._save_sync_state.saves["42"].files
-        assert "pokemon.srm" in files_state
-        assert "pokemon.sav" not in files_state
-        assert files_state["pokemon.srm"].last_sync_hash == canonical_hash
-
-    @pytest.mark.asyncio
-    async def test_resolve_filename_symmetry_use_server_keys_state_on_canonical(self, tmp_path):
-        """use_server with a mismatched frontend filename still writes at the canonical path.
-
-        Pairs with ``test_resolve_filename_symmetry_keep_local_uses_canonical_target``
-        to assert both branches converge on the same end state regardless of
-        the frontend label.
-        """
-        svc, fake = make_service(tmp_path)
-        _enable_sync_with_device(svc)
-        _install_rom(svc, tmp_path)
-        _create_save(tmp_path, content=b"local-stale")
-
-        ss = _server_save_with_syncs(
-            device_syncs=[{"device_id": "device-1", "is_current": False}],
-        )
-        ss["file_extension"] = "srm"
-        server_bytes = tmp_path / "server-content.bin"
-        server_bytes.write_bytes(b"server-truth")
-        fake.saves[100] = ss
-        fake.uploaded_files[100] = str(server_bytes)
-
-        result = await svc.resolve_sync_conflict(
-            rom_id=42,
-            filename="pokemon.sav",  # frontend label diverges from canonical
-            server_save_id=100,
-            action="use_server",
-        )
-
-        assert result["success"] is True
-        # Download landed at the canonical path, not the frontend label.
-        canonical_path = tmp_path / "saves" / "gba" / "pokemon.srm"
-        assert canonical_path.read_bytes() == b"server-truth"
-        assert not (tmp_path / "saves" / "gba" / "pokemon.sav").exists()
-        files_state = svc._save_sync_state.saves["42"].files
-        assert "pokemon.srm" in files_state
-        assert "pokemon.sav" not in files_state
-        assert files_state["pokemon.srm"].tracked_save_id == 100
-
-    @pytest.mark.asyncio
-    async def test_resolve_keep_local_raises_when_canonical_path_missing(self, tmp_path):
-        """If the local file is not at the canonical path, keep_local raises.
-
-        Defensive companion to the symmetry fix: we never silently rename
-        across extensions to satisfy a frontend label. A file named
-        ``pokemon.sav`` on disk while the server save's canonical target is
-        ``pokemon.srm`` must surface as ``FileNotFoundError`` so the user
-        can rectify the mismatch instead of having two divergent files
-        appear from a successful-looking resolve.
-        """
-        svc, fake = make_service(tmp_path)
-        _enable_sync_with_device(svc)
-        _install_rom(svc, tmp_path)
-        # Local file only at the non-canonical path.
-        saves_dir = tmp_path / "saves" / "gba"
-        saves_dir.mkdir(parents=True, exist_ok=True)
-        (saves_dir / "pokemon.sav").write_bytes(b"local-noncanonical")
-
-        ss = _server_save_with_syncs(
-            device_syncs=[{"device_id": "device-1", "is_current": False}],
-        )
-        ss["file_extension"] = "srm"
-        fake.saves[100] = ss
-
-        result = await svc.resolve_sync_conflict(
-            rom_id=42,
-            filename="pokemon.sav",
-            server_save_id=100,
-            action="keep_local",
-        )
-
-        assert result["success"] is False
-        assert "not found" in result["message"].lower()
-        # No upload was attempted.
-        assert not any(c[0] == "upload_save" for c in fake.call_log)
-
-
-class TestResolveSyncConflictStaleConflict:
-    """Round-trip validation of ``server_save_id`` guards against a third-device
-    race: another device PUTs into the slot while the conflict modal is open,
-    and the client's stale id should not be silently rewritten."""
-
-    @pytest.mark.asyncio
-    async def test_resolve_keep_local_rejects_when_server_head_advanced(self, tmp_path):
-        """Client passes server_save_id=100; server head is id=200 → stale_conflict.
-
-        Asserts the dangerous PUT never fires and state stays untouched.
-        """
-        svc, fake = make_service(tmp_path)
-        _enable_sync_with_device(svc)
-        _install_rom(svc, tmp_path)
-        _create_save(tmp_path, content=b"local-edited")
-
-        # The id=100 save the modal was opened against is no longer the head:
-        # a third device uploaded id=200 with a newer updated_at into the slot.
-        newer_server = _server_save_with_syncs(
-            save_id=200,
-            updated_at="2026-01-02T00:00:00Z",
-            device_syncs=[{"device_id": "device-2", "is_current": True}],
-        )
-        fake.saves[200] = newer_server
-
-        # Snapshot state so the no-mutation assertion is exact.
-        state_before = svc._save_sync_state.to_dict()
-
-        result = await svc.resolve_sync_conflict(
-            rom_id=42,
-            filename="pokemon.srm",
-            server_save_id=100,
-            action="keep_local",
-        )
-
-        assert result["success"] is False
-        assert result["error_code"] == "stale_conflict"
-        assert result["message"]
-        # No PUT/POST fired against the head save — the whole point of the guard.
-        assert not any(c[0] == "upload_save" for c in fake.call_log)
-        # State unchanged.
-        assert svc._save_sync_state.to_dict() == state_before
-
-    @pytest.mark.asyncio
-    async def test_resolve_use_server_rejects_when_server_head_advanced(self, tmp_path):
-        """use_server with a stale server_save_id is also rejected — the user
-        chose to download id=100, not id=200; surfacing id=200 silently would
-        also be a silent overwrite of the user's intent."""
-        svc, fake = make_service(tmp_path)
-        _enable_sync_with_device(svc)
-        _install_rom(svc, tmp_path)
-        save_path = _create_save(tmp_path, content=b"local-stale")
-
-        newer_server = _server_save_with_syncs(
-            save_id=200,
-            updated_at="2026-01-02T00:00:00Z",
-            device_syncs=[{"device_id": "device-2", "is_current": True}],
-        )
-        # Make the server's "newer" save downloadable so we can prove the
-        # download never fired.
-        server_bytes = tmp_path / "third-device-bytes.bin"
-        server_bytes.write_bytes(b"third-device-content")
-        fake.saves[200] = newer_server
-        fake.uploaded_files[200] = str(server_bytes)
-
-        # Snapshot state so the no-mutation assertion is exact.
-        state_before = svc._save_sync_state.to_dict()
-
-        result = await svc.resolve_sync_conflict(
-            rom_id=42,
-            filename="pokemon.srm",
-            server_save_id=100,
-            action="use_server",
-        )
-
-        assert result["success"] is False
-        assert result["error_code"] == "stale_conflict"
-        # Local file untouched — no silent download of the wrong server save.
-        assert save_path.read_bytes() == b"local-stale"
-        # State unchanged.
-        assert svc._save_sync_state.to_dict() == state_before
-
-    @pytest.mark.asyncio
-    async def test_resolve_succeeds_when_server_head_matches(self, tmp_path):
-        """Same flow but id=100 is still the head — resolution proceeds normally."""
-        svc, fake = make_service(tmp_path)
-        _enable_sync_with_device(svc)
-        _install_rom(svc, tmp_path)
-        save_path = _create_save(tmp_path, content=b"local-edited")
-        local_hash = _file_md5(str(save_path))
-
-        ss = _server_save_with_syncs(
-            save_id=100,
-            updated_at="2026-01-01T00:00:00Z",
-            device_syncs=[{"device_id": "device-1", "is_current": False}],
-        )
-        other = tmp_path / "server-bytes.bin"
-        other.write_bytes(b"server-flavor")
-        fake.saves[100] = ss
-        fake.uploaded_files[100] = str(other)
-
-        result = await svc.resolve_sync_conflict(
-            rom_id=42,
-            filename="pokemon.srm",
-            server_save_id=100,
-            action="keep_local",
-        )
-
-        assert result["success"] is True
-        assert result["action"] == "keep_local"
-        upload_calls = [c for c in fake.call_log if c[0] == "upload_save"]
-        assert len(upload_calls) == 1
-        assert upload_calls[0][2]["save_id"] == 100
-        file_state = svc._save_sync_state.saves["42"].files["pokemon.srm"]
-        assert file_state.last_sync_hash == local_hash
-
-
-class TestMigrationPendingGuards:
-    """The defense-in-depth migration-pending guards in pre_launch_sync and
-    post_exit_sync. The decorator on the public callable is the primary gate;
-    this in-engine guard catches a future caller that bypasses it (engine.py
-    lines 286-292 / 340-347)."""
-
-    @pytest.mark.asyncio
-    async def test_pre_launch_sync_returns_blocked_when_migration_pending(self, tmp_path):
-        """pre_launch_sync must short-circuit with blocked_by_migration=True."""
-        svc, fake = make_service(
-            tmp_path,
-            is_retrodeck_migration_pending=lambda: True,
-        )
-        svc._save_sync_state.settings.save_sync_enabled = True
-        svc._save_sync_state.device_id = "test-device"
-        _install_rom(svc, tmp_path)
-        _create_save(tmp_path, content=b"unsyncable")
-
-        result = await svc.pre_launch_sync(42)
-
-        assert result["success"] is False
-        assert result["blocked_by_migration"] is True
-        assert result["synced"] == 0
-        # No upload/download initiated — the guard fired before sync ran.
-        assert not any(c[0] in ("upload_save", "download_save_content") for c in fake.call_log)
-
-    @pytest.mark.asyncio
-    async def test_post_exit_sync_returns_blocked_when_migration_pending(self, tmp_path):
-        """post_exit_sync must short-circuit with blocked_by_migration=True."""
-        svc, fake = make_service(
-            tmp_path,
-            is_retrodeck_migration_pending=lambda: True,
-        )
-        svc._save_sync_state.settings.save_sync_enabled = True
-        svc._save_sync_state.device_id = "test-device"
-        _install_rom(svc, tmp_path)
-        _create_save(tmp_path, content=b"unsyncable")
-
-        result = await svc.post_exit_sync(42)
-
-        assert result["success"] is False
-        assert result["blocked_by_migration"] is True
-        assert result["synced"] == 0
-        assert not any(c[0] in ("upload_save", "download_save_content") for c in fake.call_log)
-
-
-class TestEnsureDeviceRegisteredFailurePaths:
-    """When register_device fails, the four sync callables must surface
-    DEVICE_NOT_REGISTERED instead of proceeding with a missing device_id
-    (engine.py lines 309-311 / 365 / 407 / 437-439)."""
-
-    @pytest.mark.asyncio
-    async def test_pre_launch_sync_returns_device_not_registered_on_failure(self, tmp_path):
-        svc, fake = make_service(tmp_path)
-        svc._save_sync_state.settings.save_sync_enabled = True
-        # No device_id set — triggers ensure_device_registered.
-        _install_rom(svc, tmp_path)
-        # register_device raises → ensure_device_registered returns success=False.
-        fake.fail_on_next(RommApiError("Server unreachable"))
-
-        result = await svc.pre_launch_sync(42)
-
-        assert result["success"] is False
-        assert "Device" in result["message"] or "device" in result["message"]
-        # No sync ran — the guard returned early.
-        assert not any(c[0] == "list_saves" for c in fake.call_log)
-
-    @pytest.mark.asyncio
-    async def test_post_exit_sync_returns_device_not_registered_on_failure(self, tmp_path):
-        svc, fake = make_service(tmp_path)
-        svc._save_sync_state.settings.save_sync_enabled = True
-        # No device_id set.
-        _install_rom(svc, tmp_path)
-        _create_save(tmp_path, content=b"data")
-        fake.fail_on_next(RommApiError("Server unreachable"))
-
-        result = await svc.post_exit_sync(42)
-
-        assert result["success"] is False
-        assert "Device" in result["message"] or "device" in result["message"]
-        # No upload ran.
-        assert not any(c[0] == "upload_save" for c in fake.call_log)
-
-    @pytest.mark.asyncio
-    async def test_sync_rom_saves_returns_device_not_registered_on_failure(self, tmp_path):
-        svc, fake = make_service(tmp_path)
-        svc._save_sync_state.settings.save_sync_enabled = True
-        # No device_id set.
-        _install_rom(svc, tmp_path)
-        fake.fail_on_next(RommApiError("Server unreachable"))
-
-        result = await svc.sync_rom_saves(42)
-
-        assert result["success"] is False
-        assert "Device" in result["message"] or "device" in result["message"]
-        assert not any(c[0] == "list_saves" for c in fake.call_log)
-
-    @pytest.mark.asyncio
-    async def test_sync_all_saves_returns_device_not_registered_on_failure(self, tmp_path):
-        svc, fake = make_service(tmp_path)
-        svc._save_sync_state.settings.save_sync_enabled = True
-        # No device_id set.
-        _install_rom(svc, tmp_path, rom_id=1, system="gba", file_name="game1.gba")
-        fake.fail_on_next(RommApiError("Server unreachable"))
-
-        result = await svc.sync_all_saves()
-
-        assert result["success"] is False
-        assert "Device" in result["message"] or "device" in result["message"]
-        # No per-ROM sync ran.
-        assert not any(c[0] == "list_saves" for c in fake.call_log)
-
-
-class TestPostExitServerOfflineGuard:
-    """post_exit_sync probes heartbeat first; on failure returns offline=True
-    instead of attempting upload (engine.py lines 358-360)."""
-
-    @pytest.mark.asyncio
-    async def test_post_exit_sync_returns_offline_when_heartbeat_raises(self, tmp_path):
-        svc, fake = make_service(tmp_path)
-        svc._save_sync_state.settings.save_sync_enabled = True
-        svc._save_sync_state.device_id = "test-device"
-        _install_rom(svc, tmp_path)
-        _create_save(tmp_path, content=b"data")
-        fake.heartbeat_raises = RommApiError("Connection refused")
-
-        result = await svc.post_exit_sync(42)
-
-        assert result["success"] is False
-        assert result["offline"] is True
-        assert result["synced"] == 0
-        # No upload was attempted after heartbeat failed.
-        assert not any(c[0] == "upload_save" for c in fake.call_log)
-
-
-class TestSyncRomSavesDisabledGuard:
-    """Public sync_rom_saves returns failure when save sync is disabled
-    (engine.py line 396)."""
-
-    @pytest.mark.asyncio
-    async def test_sync_rom_saves_disabled_returns_failure(self, tmp_path):
-        svc, fake = make_service(tmp_path)
-        # save_sync_enabled stays False by default.
-        result = await svc.sync_rom_saves(42)
-
-        assert result["success"] is False
-        assert "disabled" in result["message"].lower()
-        assert result["synced"] == 0
-        # No list_saves issued — the guard fired before sync ran.
-        assert not any(c[0] == "list_saves" for c in fake.call_log)
-
-
-class TestSyncCallableErrorMessages:
-    """The error-count clause in each public callable's success message
-    (engine.py lines 318 / 380 / 414). Driven by stubbing _sync_rom_saves
-    to return a non-empty errors list."""
-
-    @pytest.mark.asyncio
-    async def test_pre_launch_sync_message_includes_error_count(self, tmp_path):
-        svc, _ = make_service(tmp_path)
-        svc._save_sync_state.settings.save_sync_enabled = True
-        svc._save_sync_state.device_id = "test-device"
-        _install_rom(svc, tmp_path)
-
-        def stub_sync(rom_id):
-            return (0, ["pokemon.srm: bad gateway"], [])
-
-        svc._sync_engine._sync_rom_saves = stub_sync  # type: ignore[method-assign]
-
-        result = await svc.pre_launch_sync(42)
-
-        assert result["success"] is False
-        assert "1 error(s)" in result["message"]
-        assert "Downloaded" in result["message"]
-
-    @pytest.mark.asyncio
-    async def test_post_exit_sync_message_includes_error_count(self, tmp_path):
-        svc, _ = make_service(tmp_path)
-        svc._save_sync_state.settings.save_sync_enabled = True
-        svc._save_sync_state.device_id = "test-device"
-        _install_rom(svc, tmp_path)
-
-        def stub_sync(rom_id):
-            return (0, ["pokemon.srm: timeout"], [])
-
-        svc._sync_engine._sync_rom_saves = stub_sync  # type: ignore[method-assign]
-
-        result = await svc.post_exit_sync(42)
-
-        assert result["success"] is False
-        assert "1 error(s)" in result["message"]
-        assert "Uploaded" in result["message"]
-
-    @pytest.mark.asyncio
-    async def test_sync_rom_saves_message_includes_error_count(self, tmp_path):
-        svc, _ = make_service(tmp_path)
-        svc._save_sync_state.settings.save_sync_enabled = True
-        svc._save_sync_state.device_id = "test-device"
-        _install_rom(svc, tmp_path)
-
-        def stub_sync(rom_id):
-            return (0, ["pokemon.srm: 502 bad gateway"], [])
-
-        svc._sync_engine._sync_rom_saves = stub_sync  # type: ignore[method-assign]
-
-        result = await svc.sync_rom_saves(42)
-
-        assert result["success"] is False
-        assert "1 error(s)" in result["message"]
-
-
-class TestSyncEngineDelegates:
-    """Cover the thin delegate methods on SyncEngine that forward to MatrixExecutor
-    or DeviceRegistry (engine.py lines 204 / 220 / 239)."""
-
-    def test_adopt_baseline_hash_delegates_to_matrix(self, tmp_path):
-        """SyncEngine._adopt_baseline_hash writes through to the matrix's state."""
-        svc, _ = make_service(tmp_path)
-
-        svc._sync_engine._adopt_baseline_hash("42", "pokemon.srm", "deadbeef" * 4)
-
-        file_state = svc._save_sync_state.saves["42"].files["pokemon.srm"]
-        assert file_state.last_sync_hash == "deadbeef" * 4
-
-    def test_build_sync_conflict_entry_delegates_to_matrix(self, tmp_path):
-        """SyncEngine._build_sync_conflict_entry builds the same dict shape as the matrix."""
-        svc, _ = make_service(tmp_path)
-        server = _server_save(save_id=77, filename="pokemon.srm", file_size_bytes=2048)
-
-        entry = svc._sync_engine._build_sync_conflict_entry(
-            rom_id=42,
-            filename="pokemon.srm",
-            server=server,
-            local_path=None,
-            local_hash=None,
-        )
-
-        assert entry["type"] == "sync_conflict"
-        assert entry["rom_id"] == 42
-        assert entry["filename"] == "pokemon.srm"
-        assert entry["server_save_id"] == 77
-        assert entry["server_size"] == 2048
-        assert "created_at" in entry
-
-    @pytest.mark.asyncio
-    async def test_list_devices_delegates_to_device_registry(self, tmp_path):
-        """SyncEngine.list_devices forwards to DeviceRegistry.list_devices."""
-        svc, fake = make_service(tmp_path)
-        svc._save_sync_state.settings.save_sync_enabled = True
-        svc._save_sync_state.server_device_id = "device-1"
-        # Seed a registered device on the fake so list_devices returns non-empty.
-        fake._registered_devices.append({"id": "device-1", "name": "test-host"})
-
-        result = await svc._sync_engine.list_devices()
-
-        assert result["success"] is True
-        assert len(result["devices"]) == 1
-        assert result["devices"][0]["is_current_device"] is True
-
-
 class TestGetServerSaveHashNonRetryable:
     """get_server_save_hash swallows non-retryable errors and returns None
     (matrix.py line 130). The retryable-raise path (line 129) is already
@@ -2674,29 +1421,6 @@ class TestDispatchUploadDefensiveBranches:
         assert errors == []
         # No upload was attempted.
         assert not any(c[0] == "upload_save" for c in fake.call_log)
-
-
-class TestPreLaunchSaveSortGate:
-    """pre_launch_sync short-circuits when a save-sort migration is pending
-    (engine.py line 297-303)."""
-
-    @pytest.mark.asyncio
-    async def test_pre_launch_sync_returns_save_sort_changed(self, tmp_path):
-        svc, fake = make_service(tmp_path)
-        svc._save_sync_state.settings.save_sync_enabled = True
-        svc._save_sync_state.device_id = "test-device"
-        _install_rom(svc, tmp_path)
-        # Flag save-sort changed via the rom_info state path used by RomInfoService.
-        svc._state["save_sort_settings"] = {"sort_by_content": True, "sort_by_core": False}
-        svc._state["save_sort_settings_previous"] = {"sort_by_content": False, "sort_by_core": False}
-
-        result = await svc.pre_launch_sync(42)
-
-        assert result["success"] is False
-        assert result["save_sort_changed"] is True
-        assert result["synced"] == 0
-        # No sync ran.
-        assert not any(c[0] in ("upload_save", "download_save_content") for c in fake.call_log)
 
 
 class TestRecordOwnUploadNoneId:
